refactor: add common utils function for getLogoutUrl (which will read the feature flag)
