+++ src/features/amUI/systemUser/SystemUserRequestPage.tsx
@@ -16,8 +16,9 @@ import type { ProblemDetail } from './types';
 import { RightsList } from './components/RightsList/RightsList';
 import { ButtonRow } from './components/ButtonRow/ButtonRow';
 import { DelegationCheckError } from './components/DelegationCheckError/DelegationCheckError';
-import { getApiBaseUrl, getLogoutUrl } from './urlUtils';
+import { getApiBaseUrl } from './urlUtils';
 import { CreateSystemUserCheck } from './components/CreateSystemUserCheck/CreateSystemUserCheck';
+import { getLogoutUrl } from '@/resources/utils/pathUtils';
 
 export const SystemUserRequestPage = () => {
   const { t } = useTranslation();
